+++ src/test/java/com/aventstack/extentreports/ParallelMethods.java
@@ -10,12 +10,10 @@ import org.testng.annotations.Test;
 import com.aventstack.extentreports.Status;
 import com.aventstack.extentreports.common.ExtentManager;
 import com.aventstack.extentreports.common.ExtentTestManager;
-import com.aventstack.extentreports.reporter.ExtentXReporter;
 
 public class ParallelMethods {
     
     private final String filePath = getOutputFolder() + getClass().getName() + ".html";
-    private final String fileName = getClass().getSimpleName();
     
     private String getOutputFolder() {
         return "test-output/";
@@ -23,16 +21,8 @@ public class ParallelMethods {
     
     @BeforeClass
     public void beforeClass() {
-        ExtentManager.createInstance(filePath);
-        
-        ExtentXReporter extentx = new ExtentXReporter("localhost");
-        extentx.config().setProjectName("extentreports-pro");
-        extentx.config().setReportName(fileName);
-        extentx.config().setServerUrl("http://localhost:1337/");
-        
-        ExtentManager.getInstance().attachReporter(extentx);
-        
-        ExtentTestManager.setReporter(ExtentManager.getInstance());;
+        ExtentManager.createInstance(filePath);       
+        ExtentTestManager.setReporter(ExtentManager.getInstance());
     }
     
     @AfterClass
