+++ src/test/java/com/aventstack/extentreports/api/NodeMediaTest.java
@@ -10,7 +10,7 @@ import org.testng.annotations.Test;
 import com.aventstack.extentreports.Base;
 import com.aventstack.extentreports.ExtentTest;
 
-public class TestMediaTests extends Base {
+public class NodeMediaTest extends Base {
 
     private final String imgName = "img";
     private URL url = getClass().getClassLoader().getResource("1.png");
@@ -18,40 +18,43 @@ public class TestMediaTests extends Base {
     
     @Test
     public void verifyIfScreenCaptureAdded(Method method) throws IOException {
-        ExtentTest test = extent
+        ExtentTest node = extent
                 .createTest(method.getName())
+                .createNode("Child")
                 .addScreenCaptureFromPath(imgPath)
                 .fail("fail");
         
-        Assert.assertEquals(test.getModel().getScreenCaptureList().size(), 1);
+        Assert.assertEquals(node.getModel().getScreenCaptureList().size(), 1);
     }
     
     @Test
     public void verifyScreenCaptureTitle(Method method) throws IOException {
-        ExtentTest test = extent
+        ExtentTest node = extent
                 .createTest(method.getName())
+                .createNode("Child")
                 .addScreenCaptureFromPath(imgPath, imgName)
                 .fail("fail");
         
-        Assert.assertEquals(test.getModel().getScreenCaptureList().size(), 1);
-        Assert.assertEquals(test.getModel().getScreenCaptureList().get(0).getName(), imgName);
+        Assert.assertEquals(node.getModel().getScreenCaptureList().size(), 1);
+        Assert.assertEquals(node.getModel().getScreenCaptureList().get(0).getName(), imgName);
     }
     
     @Test
     public void verifyMultipleScreenCaptures(Method method) throws IOException {
         int times = 4;
         
-        ExtentTest test = extent
+        ExtentTest node = extent
                 .createTest(method.getName())
+                .createNode("Child")
                 .fail("fail");
 
         for (int ix = 0; ix < times; ix++) {
-            test.addScreenCaptureFromPath(imgPath, imgName);
+            node.addScreenCaptureFromPath(imgPath, imgName);
         }
         
-        Assert.assertEquals(test.getModel().getScreenCaptureList().size(), times);
+        Assert.assertEquals(node.getModel().getScreenCaptureList().size(), times);
         for (int ix = 0; ix < times; ix++) {
-            Assert.assertEquals(test.getModel().getScreenCaptureList().get(ix).getName(), imgName);
+            Assert.assertEquals(node.getModel().getScreenCaptureList().get(ix).getName(), imgName);
         }
     }
